backend/requirements.txt
Django==3.0.7
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
mysqlclient==1.4.6
django-cors-headers==3.2.1 
money==1.3.0
django-money==1.0
python-memcached==1.59
djangorestframework-simplejwt==4.4.0
